+++ app/helpers/work_packages_helper.rb
@@ -29,6 +29,8 @@
 #++
 
 module WorkPackagesHelper
+  include AccessibilityHelper
+
   def work_package_api_done_ratio_if_enabled(api, issue)
     if Setting.work_package_done_ratio != 'disabled'
       api.done_ratio  issue.done_ratio
@@ -199,7 +201,7 @@ module WorkPackagesHelper
     end
 
     description = if work_package.description.blank?
-                    "-"
+                    empty_element_tag
                   else
                     textilizable(description_lines.join(""))
                   end
@@ -328,20 +330,23 @@ module WorkPackagesHelper
     ].compact
   end
 
+  def work_package_show_attribute_list(work_package)
+    main_attributes = work_package_show_main_attributes(work_package)
+    custom_field_attributes = work_package_show_custom_fields(work_package)
+    hook_attributes = YAML::load(call_hook(:view_work_packages_show_details_bottom, :issue => work_package))
+
+    (main_attributes | custom_field_attributes | (hook_attributes || [])).compact
+  end
+
+  def group_work_package_attributes(attribute_list)
+    attributes = {}
+    attributes[:left], attributes[:right] = attribute_list.each_slice((attribute_list.count+1) / 2).to_a
+
+    attributes
+  end
+
   def work_package_show_attributes(work_package)
-    [
-      work_package_show_status_attribute(work_package),
-      work_package_show_start_date_attribute(work_package),
-      work_package_show_priority_attribute(work_package),
-      work_package_show_due_date_attribute(work_package),
-      work_package_show_assigned_to_attribute(work_package),
-      work_package_show_progress_attribute(work_package),
-      work_package_show_responsible_attribute(work_package),
-      work_package_show_category_attribute(work_package),
-      work_package_show_spent_time_attribute(work_package),
-      work_package_show_fixed_version_attribute(work_package),
-      work_package_show_estimated_hours_attribute(work_package)
-    ].compact
+    group_work_package_attributes work_package_show_attribute_list(work_package)
   end
 
   def work_package_show_table_row(attribute, klass = nil, &block)
@@ -357,7 +362,7 @@ module WorkPackagesHelper
     work_package_show_table_row(:status) do
       work_package.status ?
         work_package.status.name :
-        "-"
+        empty_element_tag
     end
   end
 
@@ -365,7 +370,7 @@ module WorkPackagesHelper
     work_package_show_table_row(:start_date, 'start-date') do
       work_package.start_date ?
         format_date(work_package.start_date) :
-        "-"
+        empty_element_tag
     end
   end
 
@@ -373,7 +378,7 @@ module WorkPackagesHelper
     work_package_show_table_row(:priority) do
       work_package.priority ?
         work_package.priority.name :
-        "-"
+        empty_element_tag
     end
   end
 
@@ -381,14 +386,14 @@ module WorkPackagesHelper
     work_package_show_table_row(:due_date) do
       work_package.due_date ?
         format_date(work_package.due_date) :
-        "-"
+        empty_element_tag
     end
   end
 
   def work_package_show_assigned_to_attribute(work_package)
     work_package_show_table_row(:assigned_to) do
       content = avatar(work_package.assigned_to, :size => "14").html_safe
-      content << (work_package.assigned_to ? link_to_user(work_package.assigned_to) : "-")
+      content << (work_package.assigned_to ? link_to_user(work_package.assigned_to) : empty_element_tag)
       content
     end
   end
@@ -396,7 +401,7 @@ module WorkPackagesHelper
   def work_package_show_responsible_attribute(work_package)
     work_package_show_table_row(:responsible) do
       content = avatar(work_package.responsible, :size => "14").html_safe
-      content << (work_package.responsible ? link_to_user(work_package.responsible) : "-")
+      content << (work_package.responsible ? link_to_user(work_package.responsible) : empty_element_tag)
       content
     end
   end
@@ -413,7 +418,7 @@ module WorkPackagesHelper
     work_package_show_table_row(:category) do
       work_package.category ?
         work_package.category.name :
-        '-'
+        empty_element_tag
     end
   end
 
@@ -421,7 +426,7 @@ module WorkPackagesHelper
     work_package_show_table_row(:spent_time) do
       work_package.spent_hours > 0 ?
         link_to(l_hours(work_package.spent_hours), work_package_time_entries_path(work_package)) :
-        "-"
+        empty_element_tag
     end
   end
 
@@ -429,7 +434,7 @@ module WorkPackagesHelper
     work_package_show_table_row(:fixed_version) do
       work_package.fixed_version ?
         link_to_version(work_package.fixed_version) :
-        "-"
+        empty_element_tag
     end
   end
 
@@ -437,7 +442,7 @@ module WorkPackagesHelper
     work_package_show_table_row(:estimated_hours) do
       work_package.estimated_hours ?
         l_hours(work_package.estimated_hours) :
-        "-"
+        empty_element_tag
     end
   end
 
@@ -491,7 +496,7 @@ module WorkPackagesHelper
             elsif work_package.status
               form.label(:status) + work_package.status.name
             else
-              form.label(:status) + "-"
+              form.label(:status) + empty_element_tag
             end
 
     WorkPackageAttribute.new(:status, field)
@@ -601,5 +606,31 @@ module WorkPackagesHelper
     end
 
     ret
+  end
+
+  private
+
+  def work_package_show_custom_fields(work_package)
+    work_package.custom_field_values.each_with_object([]) do |v, a|
+      a << work_package_show_table_row(v.custom_field.name, '') do
+        v.value.blank? ? empty_element_tag : simple_format_without_paragraph(h(show_value(v)))
+      end
+    end
+  end
+
+  def work_package_show_main_attributes(work_package)
+    [
+       work_package_show_status_attribute(work_package),
+       work_package_show_priority_attribute(work_package),
+       work_package_show_assigned_to_attribute(work_package),
+       work_package_show_responsible_attribute(work_package),
+       work_package_show_category_attribute(work_package),
+       work_package_show_estimated_hours_attribute(work_package),
+       work_package_show_start_date_attribute(work_package),
+       work_package_show_due_date_attribute(work_package),
+       work_package_show_progress_attribute(work_package),
+       work_package_show_spent_time_attribute(work_package),
+       work_package_show_fixed_version_attribute(work_package)
+     ]
   end
 end