+++ doc/CHANGELOG.md
@@ -45,6 +45,7 @@ See doc/COPYRIGHT.rdoc for more details.
 * `#2160` [Accessibility] Link form elements to their label - new LDAP authentication
 * `#2161` [Accessibility] Link form elements to their label - new color
 * `#2229` [Accessibility] low contrast in calendar view
+* `#2258` [Accessibility] linearisation of issue show form
 * `#2500` Change default configuration in new OpenProject application so new projects are not public by default
 * `#3370` [Design] Clean-up and refactoring existing CSS for content area
 * `#3528` [Data Migration] Type 'none' is not migrated properly in Timelines